+++ packages/components/src/components.d.ts
@@ -1226,15 +1226,15 @@ export namespace Components {
     }
     interface PTextList {
         /**
-          * The type of the text list.
+          * The type of the list.
          */
         "listType"?: ListType;
         /**
-          * The list style type of an ordered list.
+          * The list style type of ordered list. Only has effect when list type is set to 'ordered'.
          */
         "orderType"?: OrderType;
         /**
-          * Adapts the text color depending on the theme. Has no effect when "inherit" is set as color prop.
+          * Adapts the text color depending on the theme.
          */
         "theme"?: Theme;
     }
@@ -2984,15 +2984,15 @@ declare namespace LocalJSX {
     }
     interface PTextList {
         /**
-          * The type of the text list.
+          * The type of the list.
          */
         "listType"?: ListType;
         /**
-          * The list style type of an ordered list.
+          * The list style type of ordered list. Only has effect when list type is set to 'ordered'.
          */
         "orderType"?: OrderType;
         /**
-          * Adapts the text color depending on the theme. Has no effect when "inherit" is set as color prop.
+          * Adapts the text color depending on the theme.
          */
         "theme"?: Theme;
     }