+++ packages/components/src/components/button-group/button-group-styles.spec.ts
@@ -2,7 +2,7 @@ import { getComponentCss } from './button-group-styles';
 import type { ButtonGroupDirection } from './button-group-utils';
 import type { BreakpointCustomizable } from '../../types';
 
-xdescribe('getComponentCss()', () => {
+describe('getComponentCss()', () => {
   it.each<BreakpointCustomizable<ButtonGroupDirection>>([
     { base: 'column', xs: 'row' },
     'column',